# docs/openapi.yml
openapi: 3.0.0
info:
  description: A Movie Rating System API that allows users to rate movies and view their ratings.
  title: Movie Rating System API
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '1.0'
servers:
  - url: http://localhost:8081/
    description: Local development server
paths:
...
  /api/v1/movies:
    get:
      description: Get a list of all movies with optional pagination and filtering
      tags:
        - movies
      summary: Get all movies
      parameters:
        - name: limit
          in: query
          description: 'Number of movies to return (default: 20)'
          schema:
            type: integer
        - name: offset
          in: query
          description: 'Offset for pagination (default: 0)'
          schema:
            type: integer
        - name: sort_by
          in: query
          description: 'Field to sort by (default: created_at)'
          schema:
            type: string
        - name: order
          in: query
          description: 'Sort order (asc or desc, default: desc)'
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      description: Create a new movie with the provided information
      tags:
        - movies
      summary: Create a new movie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMovieRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/search/movies:
    get:
      description: Search for movies based on various criteria
      tags:
        - movies
      summary: Search movies
      parameters:
        - name: query
          in: query
          description: Search query
          schema:
            type: string
        - name: genre
          in: query
          description: Movie genre
          schema:
            type: string
        - name: director
          in: query
          description: Movie director
          schema:
            type: string
        - name: min_year
          in: query
          description: Minimum release year
          schema:
            type: integer
        - name: max_year
          in: query
          description: Maximum release year
          schema:
            type: integer
        - name: limit
          in: query
          description: 'Number of movies to return (default: 20)'
          schema:
            type: integer
        - name: offset
          in: query
          description: 'Offset for pagination (default: 0)'
          schema:
            type: integer
        - name: sort_by
          in: query
          description: 'Field to sort by (default: created_at)'
          schema:
            type: string
        - name: order
          in: query
          description: 'Sort order (asc or desc, default: desc)'
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchMoviesResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/search/movies/{id}:
    get:
      description: Get detailed information about a specific movie
      tags:
        - movies
      summary: Get a movie by ID
      parameters:
        - name: id
          in: path
          required: true
          description: Movie ID
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateMovieRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        director:
          type: string
        genre:
          type: string
        release_year:
          type: integer
        duration_mins:
          type: integer
        language:
          type: string
        country:
          type: string
        budget:
          type: integer
        revenue:
          type: integer
        rating:
          type: string
        imdb_id:
          type: string
        poster_url:
          type: string
    MovieResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        director:
          type: string
        genre:
          type: string
        release_year:
          type: integer
        duration_mins:
          type: integer
        language:
          type: string
        country:
          type: string
        budget:
          type: integer
        revenue:
          type: integer
        rating:
          type: string
        imdb_id:
          type: string
        poster_url:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
    SearchMoviesResponse:
      type: object
      properties:
        movies:
          type: array
          items:
            $ref: '#/components/schemas/MovieResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
        query:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic